+++ CHANGES.md
@@ -7,6 +7,8 @@
 - Having multiple keys with the same name should no longer completely break the build
 - Maven dependency resolver now checks `WEMI_MAVEN_OS_NAME`, `WEMI_MAVEN_OS_ARCH`, `WEMI_MAVEN_OS_VERSION` and `WEMI_MAVEN_OS_FAMILY`
     environment variables to detect relevant profile activation properties
+- Added a new `Dependency` type, `TypeChooseByPackaging`, which allows Wemi to choose the artifact automatically
+- Added `SystemInfo` for all operating system and processor architecture detection needs 
 
 # 0.13 2020-06-23
 - Add utility functions for source file and classpath generation, see the `wemi.generation` package